\docType{package}
\name{FPF-package}
\alias{FPF}
\alias{FPF-package}
\title{fMRI Prevalence Finder}
\description{
  Fit voxelwise mixture models in fMRI group studies.
}
\details{
  \tabular{ll}{ Package: \tab rosenblatt1\cr Type: \tab
  Package\cr Version: \tab 0.5\cr Date: \tab 2012-08-19\cr
  License: \tab GPL (>= 2)\cr }

  Mixture model fitting for random effect in fMRI group
  studies.
}
\author{
  Jonathan Rosenblatt<john.ros@gmail.com>
}

