\docType{package}
\name{rosenblatt1-package}
\alias{rosenblatt1}
\alias{rosenblatt1-package}
\title{Fit mixture models to fMRI group studies}
\description{
  Fit mixture models to fMRI group studies.
}
\details{
  \tabular{ll}{ Package: \tab rosenblatt1\cr Type: \tab
  Package\cr Version: \tab 0.5\cr Date: \tab 2012-08-19\cr
  License: \tab GPL (>= 2)\cr }

  Mixture model fitting for random effect in fMRI group
  studies.
}
\author{
  Jonathan Rosenblatt
}

